Rename ixmp.utils → ixmp.util
- Consistent with the rest of the message_ix stack.
- Add to DeprecatedPathFinder.
- Adjust references to satisfy mypy.
